- Implement update movie

+++ Movies.Contracts/IMovieGrainClient.cs
@@ -6,6 +6,6 @@ namespace Movies.Contracts
 	public interface IMovieGrainClient
 	{
 		Task<MovieModel> Get(long id);
-		Task Set(int id, string name, string description, string img, string key, string length, decimal rate);
+		Task Set(long id, string name, string description, string img, string key, string length, decimal rate);
 	}
 }
